+++ errors/errors.go
@@ -1,7 +1,7 @@
 package errors
 
 type BaseError struct {
-	Code    uint8  `json:"code"`
+	Code    uint   `json:"code"`
 	Message string `json:"message"`
 	Data    string `json:"data"`
 }
